Move inference output types to tensorzero-inference-types
This required making a few inherent methods into freestanding methods.
However, it brings us closer to removing the
'ts-executor-pool -> tensorzero-core' dependency edge
