+++ crates/tensorzero-core/src/endpoints/datasets/v1/conversion_utils.rs
@@ -8,7 +8,9 @@ use crate::endpoints::datasets::v1::types::{
 };
 use crate::error::{Error, ErrorDetails};
 use crate::function::FunctionConfig;
-use crate::inference::types::{FetchContext, InputExt, JsonInferenceOutput};
+use crate::inference::types::{
+    FetchContext, InputExt, JsonInferenceOutput, validate_content_block_chat_output,
+};
 use crate::jsonschema_util::JSONSchema;
 use crate::tool::ToolCallConfigDatabaseInsert;
 
@@ -46,7 +48,7 @@ impl CreateChatDatapointRequest {
         let validated_output = if let Some(output) = self.output {
             let validation_futures = output
                 .into_iter()
-                .map(|output| output.into_validated(tool_config.as_ref()));
+                .map(|output| validate_content_block_chat_output(output, tool_config.as_ref()));
             let validated_output = join_all(validation_futures).await;
             Some(validated_output)
         } else {
