+++ crates/ts-executor-pool/Cargo.toml
@@ -14,6 +14,7 @@ regex.workspace = true
 tokio.workspace = true
 tokio-util.workspace = true
 tensorzero-core.workspace = true
+tensorzero-inference-types = { path = "../tensorzero-inference-types" }
 tensorzero-ts-types = { path = "../tensorzero-ts-types" }
 tensorzero-types = { path = "../tensorzero-types" }
 deno_error = "0.7"
